pyfaf: Added new variable into solution finder
I've added missing variables (title, type, certainty) into solution finder.
For PrefilterSolutionFinder is certainty set on 100 because we don't
want notify about this reports.This value represent BINGO* constant.

I've added certainty into ProbableFixSolutionFinder and set on 99. This
is because Probable fix is a special case of solution about which we
want sending notification to maintainers. The value 99 represent constant
ALMOST_THE_SAME*.

* https://github.com/abrt/faf/wiki/Improved-solutions-protocol

Signed-off-by: Patrik Helia <phelia@redhat.com>