webfaf: unique data in report detail, import optimalization
Current state:
    In current state we are showing data from first occurence in graph.
This causes the graphis are unreadable.

New state:
    In new version are showing original data by default and unique
data are hide (prepare for future).
If we choosed big history range then graphs was unreadable, so we choose
20 for all (20 days for daily grapsh, 20 weeks for weekly and 20 month for
monthly). Also if we don't have data for this
date period then is settuped count on 0 for maximal range.

In html was updated JS script for prepare data and drawing graphs.

Also i made import optimalization.

Signed-off-by: Patrik Helia <phelia@redhat.com>